fix: move logic to viewmodel
