+++ app/src/main/java/org/fossasia/openevent/general/aboutEvent/AboutEventActivity.kt
@@ -11,10 +11,7 @@ import android.widget.Toast
 import kotlinx.android.synthetic.main.fragment_about_event.*
 import org.fossasia.openevent.general.R
 import org.fossasia.openevent.general.event.Event
-import org.fossasia.openevent.general.event.EventUtils
-import org.fossasia.openevent.general.ticket.AboutEventViewModel
 import org.koin.android.architecture.ext.viewModel
-import timber.log.Timber
 import java.lang.StringBuilder
 
 class AboutEventActivity : AppCompatActivity(), AppBarLayout.OnOffsetChangedListener {
@@ -60,9 +57,9 @@ class AboutEventActivity : AppCompatActivity(), AppBarLayout.OnOffsetChangedList
         aboutEventContent.text = event.description
         val dateString = StringBuilder("")
 
-        aboutEventDetails.text = dateString.append(getAboutEventFormattedDate(event.startsAt))
+        aboutEventDetails.text = dateString.append(aboutEventViewModel.getAboutEventFormattedDate(event.startsAt))
                                            .append(" - ")
-                                           .append(getAboutEventFormattedDate(event.endsAt))
+                                           .append(aboutEventViewModel.getAboutEventFormattedDate(event.endsAt))
 
         eventName.text = event.name
     }
@@ -76,7 +73,6 @@ class AboutEventActivity : AppCompatActivity(), AppBarLayout.OnOffsetChangedList
 
         val maxScroll = appBarLayout.totalScrollRange
 
-        Timber.d("MAX "+maxScroll)
         val percentage =  Math.abs(verticalOffset).toFloat() / maxScroll.toFloat()
 
         if (percentage == 1f && isHideToolbarView) {
@@ -84,17 +80,14 @@ class AboutEventActivity : AppCompatActivity(), AppBarLayout.OnOffsetChangedList
             if (TextUtils.isEmpty(eventExtra.name)) {
                 detailsHeader.visibility = View.GONE
                 aboutEventCollapsingLayout.title = eventExtra.name
-                Timber.d("collapse "+eventExtra.name)
                 isHideToolbarView = !isHideToolbarView
             } else {
                 detailsHeader.visibility = View.GONE
                 aboutEventCollapsingLayout.title = eventExtra.name
-                Timber.d("collapse 2"+eventExtra.name)
                 isHideToolbarView = !isHideToolbarView
             }
         } else if (percentage < 1f && !isHideToolbarView) {
             //Not Collapsed
-            Timber.d("no collapse"+eventExtra.name)
             detailsHeader.visibility = View.VISIBLE
             aboutEventCollapsingLayout.title = " "
             isHideToolbarView = !isHideToolbarView
@@ -109,22 +102,5 @@ class AboutEventActivity : AppCompatActivity(), AppBarLayout.OnOffsetChangedList
             }
             else -> super.onOptionsItemSelected(item)
         }
-    }
-
-    private fun getAboutEventFormattedDate(date: String): String{
-
-        val dateString = EventUtils.getLocalizedDateTime(date)
-
-        //Format Month
-        val month = dateString.dayOfWeek
-        val lowerCaseMonth = month.toString().toLowerCase()
-        val formatMonth = (lowerCaseMonth.substring(0, 1).toUpperCase() + lowerCaseMonth.substring(1)).substring(0, 3)
-
-        //Format Day
-        val day = dateString.month
-        val lowerCaseDay = day.toString().toLowerCase()
-        val formatDay = (lowerCaseDay.substring(0, 1).toUpperCase() + lowerCaseDay.substring(1)).substring(0, 3)
-
-        return formatMonth + ", " + formatDay + " " + dateString.dayOfMonth
     }
 }
