+++ app/src/main/java/org/fossasia/openevent/general/di/Modules.kt
@@ -9,6 +9,7 @@ import okhttp3.Authenticator
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.fossasia.openevent.general.OpenEventDatabase
+import org.fossasia.openevent.general.aboutEvent.AboutEventViewModel
 import org.fossasia.openevent.general.auth.*
 import org.fossasia.openevent.general.data.Preference
 import org.fossasia.openevent.general.event.*
